modified:   subprojects/toolkit/main.cpp

+++ subprojects/toolkit/main.cpp
@@ -58,7 +58,7 @@ int main(int argc, char *argv[])
   }
 
   if (application.arguments().count())
-    id = application.arguments().at(0);
+    id = application.arguments().at(0).toInt();
 
   MainWindow window(&interface, id);
   QObject::connect(&interface, SIGNAL(dbusClassicControllerPlugged(quint32)), &window, SLOT(dbusClassicPlugged(quint32)));
